For circuit schematic V3.0 (version v2.1) 2019.12.27 The testing team gave some bugs. Fix these bugs.

+++ driver/adcdrv.c
@@ -61,7 +61,7 @@ int get_AD(uint32_t adchl)
 	sample = ADC_DataSingleGet(ADC0);
 
 	// Calculate input voltage in mV
-	temp = ((float)sample * 2500) / 4096;
+	temp = (int)(((float)sample * 2500) / 4096);
 
 	return temp;
 }
